Replace usage of QString::asprintf for new Qt
Qt5 removed the instance method QString::asprintf, and later replaced
it with a static method of the same name, causing CoreHost to compile,
but pass empty strings to the managed entry point.

The static QString::number method is available in both Qt4 and Qt5,
and as a bonus the overload should resolve to the appropriate pointer
length for the target.